merge of master

+++ BT-Tracking/AppDelegate.swift
@@ -13,6 +13,7 @@ import FirebaseAuth
 import FirebaseFunctions
 #endif
 import RealmSwift
+import RxSwift
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -26,6 +27,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     private(set) static var inBackground: Bool = false
+    private let bag = DisposeBag()
 
     // MARK: - Globals
 
@@ -38,7 +40,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     lazy var scannerStore: ScannerStore = {
         let store = ScannerStore(
             scanningPeriod: RemoteValues.collectionSeconds,
-            dataPurgeInterval: RemoteValues.persistDataDays
+            dataPurgeInterval: RemoteValues.persistDataInterval
         )
         AppDelegate.shared.scanner.add(delegate: store)
         return store
@@ -80,7 +82,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         )
 
         Realm.Configuration.defaultConfiguration = configuration
-        scannerStore.deleteOldRecordsIfNeeded()
+
+        if Auth.isLoggedIn {
+            scannerStore.deleteOldRecordsIfNeeded()
+        }
     }
 
     private func setupInterface() {
@@ -92,7 +97,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         let storyboard: UIStoryboard
         #if !targetEnvironment(macCatalyst)
 
-        if Auth.auth().currentUser == nil || AppSettings.BUID == nil {
+        if !Auth.isLoggedIn {
             try? Auth.auth().signOut()
             storyboard = UIStoryboard(name: "Signup", bundle: nil)
         } else {
@@ -106,6 +111,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window.rootViewController = storyboard.instantiateInitialViewController()
     }
     
+    private func setupBackgroundMode(for application: UIApplication) {
+        application.setMinimumBackgroundFetchInterval(UIApplication.backgroundFetchIntervalMinimum)
+    }
+    
     // MARK: - UIApplicationDelegate
 
     var window: UIWindow? = nil
@@ -115,6 +124,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         generalSetup()
         setupInterface()
+        setupBackgroundMode(for: application)
         
         return true
     }
@@ -127,8 +137,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             backgroundTask = .invalid
         }
         inBackgroundStage = false
-        
-        scannerStore.deleteOldRecordsIfNeeded()
+
+        if Auth.isLoggedIn {
+            scannerStore.deleteOldRecordsIfNeeded()
+        }
     }
 
     func applicationWillResignActive(_ application: UIApplication) {
@@ -204,6 +216,45 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             completionHandler(.noData)
         } else {
             completionHandler(.noData)
+        }
+    }
+    
+    // MARK: - Background fetch
+    
+    func application(_ application: UIApplication, performFetchWithCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void) {
+        backgroundTask = application.beginBackgroundTask (expirationHandler: { [unowned self] in
+            log("AppDelegate background: Background task expired")
+            application.endBackgroundTask(self.backgroundTask)
+            self.backgroundTask = .invalid
+        })
+        fetchRemoteConfig()
+            .subscribe(onSuccess: { [unowned self] _ in
+                log("AppDelegate background: Remote config updated")
+                completionHandler(.newData)
+                application.endBackgroundTask(self.backgroundTask)
+                self.backgroundTask = .invalid
+            }, onError: { [unowned self] error in
+                log("AppDelegate background: Remote config error")
+                completionHandler(.failed)
+                application.endBackgroundTask(self.backgroundTask)
+                self.backgroundTask = .invalid
+            })
+            .disposed(by: bag)
+    }
+    
+    private func fetchRemoteConfig() -> Single<Void> {
+        return Single<Void>.create { single in
+            RemoteConfig.remoteConfig().fetch(withExpirationDuration: 1800) { _, error in
+                if let error = error {
+                    log("AppDelegate background: Got an error fetching remote values \(error)")
+                    single(.error(error))
+                    return
+                }
+                RemoteConfig.remoteConfig().activate()
+                log("AppDelegate background: Retrieved values from the Firebase Remote Config!")
+                single(.success(()))
+            }
+            return Disposables.create()
         }
     }
 }
